Simplified animation, refactoring controlls
Removed Animation class. Instead, raw animation data is stored in a
constant struct. The AnimationComponent is responsible for updating
itself.

Started to refactor input systems. Actors now have a pointer to their
control method. ATM only controllers are (bareley) implemented. In the
future, AI should be an option.

Made animations able to not be looped, and set them to not looping by
default.