Application of shapes, other stuff
Changed Collision to use Shapes.h and its helpers.

Implemented a force out script node, so that actors that collide can
stop each other.

Made Collision Bounds on an actor actually represent the visual square
of the actor's sprite.

Added an IsDead() method for cleanup

Added subtraction and += -= operators for Vector types

Added PoolItemPtr to ObjectPool - a semi-smart pointer that keeps track
of where in the pool a bit of data came from. Inserting into a pool now
returns a PoolItemPtr. This is useful for removals, as passing in a
PoolItemPtr will resultin O(1) removals and will still allow for easy
access to stored data

Decided that I go too long between commits and need to do them whenever
I finish an individual task.