rough dungeon room linking idea
Roghed out the core idea for the dungeon creation at an object level,
this might be able to create an interesting but non sensical dungeon.
First pass on te concept will create a more logical layout in the future
- this would only work in discrete rooms, zelda style.  Will keep
iterating to create something more viable for our game, .. but yay some
ideas!