Plugged in Toli's BG
This was actually a really good test and taught us a few things:

1 - If you write a Sprite Def without an animation reference, the
resulting animation will just be static and won't cause any problems.
It's a shame that we have to waste a few ints for state, but that's not
so pivotal atm.

2 -We really need to make resource loading a step at load time, rather
than having every sprite hard-coded. It'll make adding sprites not be a
PITA for anyone not named Phil.

3 - This flat directory is AWWFULLLLLLLL and really needs to be
organized.